ES-5577 Replaced Long with Int for minimumRenewals, maximumRenewals

+++ src/main/kotlin/com/procurement/access/application/model/params/ValidateLotsDataForDivisionParams.kt
@@ -97,8 +97,8 @@ data class ValidateLotsDataForDivisionParams(
 
             data class Renewal(
                 val description: String?,
-                val minimumRenewals: Long?,
-                val maximumRenewals: Long?,
+                val minimumRenewals: Int?,
+                val maximumRenewals: Int?,
                 val period: Period?
             ) {
                 data class Period(
